Cooldown Display & More
The HUD now has a cooldown display which shows all of the attack's cooldowns. The icons also change depending on the percentage of the cooldown.
Updated the lightmap
Added a message when the player tries to attack with no stamina
Fixed an issue where the player would get stuck in a loop when attacking with no stamina
Fixed an issue where changing class and upgrading health or damage would not charge the player

To Do:
Sliding
Enemy Damage when sliding
Footstep SFX
Monster SFX
Enemy Spawn VFX